\name{readKNMI}
\alias{readKNMI}
\alias{KNMIstations}

\title{
  Reads data files from the Dutch governmental organisation KNMI. 
}

\description{
  \code{readKNMI} reads text files with weather data generated by the Dutch governmental organisation KNMI.   (met office, KONINKLIJK NEDERLANDS METEOROLOGISCH INSTITUUT).

  The dutch is translated into english, and the units are converted to more standard units

  \code{KNMIstations} provides information about the stations, i.e. the type, height, coordinates, ...
}

\usage{
readKNMI (file, dir="", attr=NULL, ...)

data(KNMIstations)
}

\arguments{
  \item{file }{The name of a the file to read. This name will be kept as an attribute of the output. 
  }
  \item{dir }{The name of the directory holding the file. 
  }
  \item{attr }{A list with information to be added to the attributes of the output.
  }
  \item{... }{Any other arguments passed to the reading function , i.e. dec, quote, fill, etc.
  }
}
\note{

A small example files in in the package's "extdata" directory: 

\itemize{
 \item for KNMI data: "KNMIsmall.txt"
}

The package directory can be grasped with function \code{system.file}. See examples.

}
\note{
for KNMI data, the coordinates of where the data came from is not kept in the file itself. 
As this information is not easily found; it has been compiled in dataset KNMIstations.
}

\value{
  Function \code{readKNMI} returns a data.frame with the data in wide format; atmospheric conditions as columns.
  
  It has the following columns:

\itemize{
 \item station: the identifier of the station (nr)
 \item datetime: the date and time of the sampling
 \item windspeed, temperature, dewpoint, radiation, pressure, humidity: the weather data
 \item value: the measured value
 \item unit: the unit of the measurement
}
  
The attributes of the object will contain: 
\itemize{
 \item source: the name of the originating organization, 
 \item file: the name of the file,
 \item processing: text keeping track of processing of the data,
 \item variables: names, units, methods, sensor used, 
 \item stations: the long names, acronyms, and position of the stationsm 
 \item EPSG: the coordinate system (EPSG nr) of the positions,
 \item processing: when the dataset was generated.
}


The attributes of the object will contain: 
\itemize{
 \item source: the name of the originating organization, 
 \item file: the name of the file
 \item processing: text keeping track of processing of the data
 \item variables: names, units, and original column of the variables
}

The attributes of the object also contain all information passed with argument "attr".
}

\examples{

# path of the external data of the package
fpath <- system.file("extdata", package="dtRtools")

dir(fpath)  # shows all files 

## =============================================================
## Read the KNMI file
## =============================================================

KNMI <- readKNMI(file = "KNMIsmall.txt", 
                 dir  = fpath)

head(KNMI, n=2)  # note that radiation is absent here
with(KNMI, 
   plot(datetime, 
        windspeed, 
        type = "l"))

AF1 <- meta(KNMI)
AF1$variables
AF1$processing


## =============================================================
## Read the files, add extra information to table and attributes
## =============================================================

KNMI2 <- readKNMI(file = "KNMIsmall.txt", 
                  dir  = fpath,
                  attr = list(author = "karline soetaert", project = "LTER-life"))

AF2 <- meta(KNMI2)
AF2$author

}
\references{
http://www.knmi.nl/
}
\seealso{
\link{readRWS}, \link{readBathymetry} for other reading functions.
}
\author{
  Karline Soetaert <karline.soetaert@nioz.nl>
}
\details{
The columns of the raw KNMI files mean (taken from the first part of the text files):

(note: most of the units have been converted in the derived data).

 \itemize{
  \item YYYYMMDD  = date (YYYY=year,MM=month,DD=day)
  \item HH        = time (HH=hour)
  \item DD        = Mean wind direction (in degrees) for the 10-minute period preceding the observation time stamp (360=north, 90=east, 180=south, 270=west, 0=calm, 990=variable)
  \item FH        = Mean wind speed (in 0.1 m/s) for the hour preceding the observation time stamp
  \item FF        = Mean wind speed (in 0.1 m/s) for the 10-minute period preceding the observation time stamp
  \item FX        = Maximum wind gust (3 second mean wind speed; in 0.1 m/s) in the preceding hour
  \item T         = Temperature (in 0.1 degrees Celsius) at 1.50 m at the time of observation
  \item TD        = Dew point temperature (in 0.1 degrees Celsius) at 1.50 m at the time of observation
  \item SQ        = Sunshine duration (in 0.1 hour) during the hourly division, calculated from global radiation (-1 for <0.05 hour)
  \item Q         = Global radiation (in J/cm2) during the hourly division
  \item DR        = Duur van de neerslag (in 0.1 uur) per uurvak / Precipitation duration (in 0.1 hour) during the hourly division
  \item RH        = Hourly precipitation amount (in 0.1 mm) (-1 for <0.05 mm)
  \item P         = Air pressure (in 0.1 hPa) reduced to mean sea level, at the time of observation
  \item VV        = Horizontal visibility at the time of observation (0=less than 100m, 1=100-200m, 2=200-300m,..., 49=4900-5000m, 50=5-6km, 56=6-7km, 57=7-8km, ..., 79=29-30km, 80=30-35km, 81=35-40km,..., 89=more than 70km)
  \item N         = Cloud cover (in octants), at the time of observation (9=sky invisible)
  \item U         = Relative atmospheric humidity (in percents) at 1.50 m at the time of observation
  \item WW        = Present weather code (00-99), description for the hourly division.
  \item IX        = Indicator present weather code (1=manned and recorded (using code from visual observations), 2,3=manned and omitted (no significant weather phenomenon to report, not available), 4=automatically recorded (using code from visual observations), 5,6=automatically omitted (no significant weather phenomenon to report, not available), 7=automatically set (using code from automated observations)
  \item M         = Fog 0=no occurrence, 1=occurred during the preceding hour and/or at the time of observation
  \item R         = Rainfall 0=no occurrence, 1=occurred during the preceding hour and/or at the time of observation
  \item S         = Snow 0=no occurrence, 1=occurred during the preceding hour and/or at the time of observation
  \item O         = Thunder  0=no occurrence, 1=occurred during the preceding hour and/or at the time of observation
  \item Y         = Ice formation 0=no occurrence, 1=occurred during the preceding hour and/or at the time of observation
  \item TZ        = Sea surface temperature (in 0.1 degrees Celsius) at the time of observation
 }
}
\keyword{utilities}
