\name{Wad_weather}
\alias{Wad_weather}
\title{
  Weather time series from the Dutch Meteorological Society (KNMI), hourly data for 2021
}

\usage{
data(Wad_weather)
}


\value{
  a data.frame of class \code{dtLife} with the following items: 
\itemize{
  \item \code{station}, Nr of station - see attributes(Wad_weather)$stations
  \item \code{datetime}, Date and time of the measurement, POSIXct format
  \item \code{windspeed}, wind velocity, 10 m above the ground, [m/s]
  \item \code{temperature}, air temperature [dg C]
  \item \code{radiation}, Light intensity, [J/m2/s]
  }
  
 and the following attributes:
\itemize{
  \item \code{stations}, Description of the stations, with their position
  \item \code{variables}, Description of the measured variables, and their units
  \item \code{datasource}, Data originator
  \item \code{file}, name of the files from which the data came
  \item \code{format}, wide format
  }  
}

\description{
 Dataset obtained from the Dutch Meteorological society. 
 
 Downloaded from https://www.knmi.nl/nederland-nu/klimatologie/uurgegevens
 
 stations:
 \itemize{
   \item 208 ("AWG-1"), 
   \item 235 ("DE KOOY VK"), 
   \item 242 ("VLIELAND"),
   \item 251 ("TERSCHELLING HOORN AWS"), 
   \item 270 ("LEEUWARDEN"), 
   \item 277 ("LAUWERSOOG AWS").
 }
 The data were generated by using \link{readKNMI}, running 

\code{
 stn_land <- c(208, 235, 242, 251, 270, 277)
 
 files <- paste("uurgeg", stn_land, "2021-2030.txt", sep="_")
 
 WeatherData <- readKNMI(dir=dir, file=files) 
 
 Wad_weather <- subset(WeatherData, 
                      subset = datetime < "2022-01-01")
                      
  Wad_weather <- Wad_weather[, c("station",  "datetime", 
                                 "windspeed", "temperature",  "radiation")]

}
}

\examples{
# ---------------------------------------------------
# Show the position of the data
# ---------------------------------------------------

Stations <- attributes(Wad_weather)$stations
Stations

# ---------------------------------------------------
# Plot the position of the data
# ---------------------------------------------------

asp <- 1/cos((mean(Stations$latitude)*pi)/180)

plot(x    = Stations$longitude, 
     y    = Stations$latitude, 
     main = "Position of weather stations", 
     xlab = "dgE", ylab = "dgN",
     pch  = 18, cex = 3, asp = asp)
text(x      = Stations$longitude, 
     y      = Stations$latitude, 
     labels = Stations$station, 
     col    = "white", cex = 0.5)
plot(Shape, add = TRUE)

# ---------------------------------------------------
# Summaries of the data
# ---------------------------------------------------
aggregate(x   = Wad_weather$windspeed, 
          by  = list(Wad_weather$station), 
          FUN = mean, na.rm=TRUE)
aggregate(x   = Wad_weather$temperature, 
          by  = list(Wad_weather$station), 
          FUN = mean, na.rm=TRUE)
aggregate(x   = Wad_weather$radiation, 
          by  = list(Wad_weather$station), 
          FUN = mean, na.rm=TRUE)

# ---------------------------------------------------
# Plot the hourly data of station 235
# ---------------------------------------------------

mf <- par(mfrow=c(3,1))
St_235 <- subset(Wad_weather, subset=station==235)

with(St_235, 
   plot(x    = datetime, 
        y    = windspeed, 
        main ="wind speed", 
        ylab = "m/s", xlab = "time",
        las  = 1, type= "l"))
with(St_235, 
   plot(x    = datetime, 
        y    = temperature, 
        main = "air temperature", 
        ylab = "m/s", xlab = "time",
        las  = 1, type = "l"))
with(St_235, 
   plot(x    = datetime, 
        y    = radiation, 
        main = "light intensity", 
        ylab = "W/m2", xlab = "time",
        las  = 1, type =  "l"))

# ---------------------------------------------------
# Plot daily-averaged data
# ---------------------------------------------------

# estimate the daily-averages

Wad_weatherDay <- average_timeseries(
            input    = Wad_weather,
            avgOver  = "day",
            datetime = "datetime",
            value    = c("windspeed", "temperature", "radiation"),
            na.rm    = TRUE
          )

mf <- par(mfrow=c(3,1))

plot(x    = Wad_weatherDay$datetime,
     y    = Wad_weatherDay$windspeed, 
     main = "daily mean wind speed", 
     ylab = "m/s", xlab = "time",
     las  = 1, type = "l")
        
plot(x    = Wad_weatherDay$datetime,
     y    = Wad_weatherDay$temperature, 
     main = "daily mean air temperature", 
     ylab = "m/s", xlab = "time",
     las = 1, type = "l")
     
plot(x    = Wad_weatherDay$datetime,
     y    = Wad_weatherDay$radiation, 
     main = "daily mean light intensity", 
     ylab = "W/m2", xlab = "time",
     las  = 1, type = "l")

par(mfrow=mf)    

\dontrun{
plot(Wad_weatherDay, type="l")
}
}
\seealso{
  \code{\link{Marsdiep}} bathymetric data.
  \code{\link{Sediment}} data set with sediment composition.
  \code{\link{Wad_watertemp_LR}} data set with (low resolution) water temperature from the Waddensea.
}
\author{
  Karline Soetaert <karline.soetaert@nioz.nl>
}
\keyword{dataset}

